Ready to install Native-Base and test Rest API calls
